Sign method does not hash empty request body
If a client sends a request with a non-nil but empty body then the Sign method (on APISigner) hashes the empty body, but the empty body is stripped from the request server-side so the server does not hash the empty body (this results in a different sign-code). This commit prevents the Sign method from hashing an empty body. Elara: TP#285005